+++ laravel-auth-exo2/app/Http/Controllers/ServiceController.php
@@ -25,6 +25,7 @@ class ServiceController extends Controller
      */
     public function create()
     {
+        $this->authorize("service-create", Service::class);
         return view("backoffice.service.create");
     }
 
@@ -36,10 +37,13 @@ class ServiceController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create", Service::class);
+
         $service = new Service();
         $service->title = $request->title;
         $service->text = $request->text;
         $service->icon = $request->icon;
+        $service->color = $request->color;
         $service->shape = $request->shape;
 
         $service->save();
@@ -67,7 +71,9 @@ class ServiceController extends Controller
      */
     public function edit(Service $service)
     {
-        return view("backoffice.service.edit", compact("service"));
+        $this->authorize("service-edit", $service);
+        $services = Service::all();
+        return view("backoffice.service.edit", compact("service", "services"));
     }
 
     /**
@@ -79,6 +85,8 @@ class ServiceController extends Controller
      */
     public function update(Request $request, Service $service)
     {
+        $this->authorize("update", $service);
+
         $service->title = $request->title;
         $service->text = $request->text;
         $service->icon = $request->icon;
@@ -97,6 +105,7 @@ class ServiceController extends Controller
      */
     public function destroy(Service $service)
     {
+        $this->authorize("delete", $service);
         $service->delete();
         return redirect()->back()->with("successMessage", "Votre service à bien été suprimmé");
     }
