+++ laravel-auth-exo2/app/Http/Controllers/TemoignageController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Service;
 use App\Models\Temoignage;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -26,6 +27,9 @@ class TemoignageController extends Controller
      */
     public function create()
     {
+        $this->authorize("temoignage", Temoignage::class);
+        
+
         return view("backoffice.temoignage.create");
     }
 
@@ -37,6 +41,8 @@ class TemoignageController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create", Service::class);
+
         $temoignage = new Temoignage();
 
         $temoignage->text = $request->text;
@@ -44,7 +50,8 @@ class TemoignageController extends Controller
         $temoignage->position = $request->position;
         $temoignage->photo = $request->file('photo')->hashName();
         $request->file('photo')->storePublicly("img/testimonials","public");
-
+        $temoignage->save();
+        
         return redirect()->route("temoignage.index")->with("successMessage", "Le temoignage de $temoignage->author à bien été ajouté");
 
 
@@ -70,6 +77,7 @@ class TemoignageController extends Controller
      */
     public function edit(Temoignage $temoignage)
     {
+        $this->authorize("temoignage", $temoignage);
         return view("backoffice.temoignage.edit", compact("temoignage"));
     }
 
@@ -82,14 +90,17 @@ class TemoignageController extends Controller
      */
     public function update(Request $request, Temoignage $temoignage)
     {
+        $this->authorize("update", $temoignage);
+
         $temoignage->text = $request->text;
         $temoignage->author = $request->author;
         $temoignage->position = $request->position;
 
-        if($request->file('logo')!= null){
+        if($request->file('photo')!= null){
             $temoignage->photo = $request->file('photo')->hashName();
             $request->file('photo')->storePublicly("img/testimonials","public");
         }
+        $temoignage->save();
         return redirect()->route("temoignage.index")->with("successMessage", "Le temoignage de $temoignage->author à bien été modifié");
 
     }
@@ -102,6 +113,8 @@ class TemoignageController extends Controller
      */
     public function destroy(Temoignage $temoignage)
     {
+        $this->authorize("delete", $temoignage);
+
         Storage::disk('public')->delete("img/testimonials" . $temoignage->photo);
         $temoignage->delete();
